Fix(Java): Add TODO comment for new severity logic and abstract it to the BlackDuckCollector to be used in all cases
